+++ src/common/classes/vector.h
@@ -24,7 +24,7 @@
  *  Contributor(s): ______________________________________.
  *
  *
- *  $Id: vector.h,v 1.10 2004-06-30 01:26:06 skidder Exp $
+ *  $Id: vector.h,v 1.11 2004-07-16 23:06:04 skidder Exp $
  *
  */
  
@@ -37,34 +37,34 @@
 namespace Firebird {
 
 // Very fast static array of simple types
-template <typename T, int Capacity>
+template <typename T, size_t Capacity>
 class Vector {
 public:
 	Vector() : count(0) {}
 	void clear() { count = 0; };
-	T& operator[](int index) {
+	T& operator[](size_t index) {
   		fb_assert(index >= 0 && index < count);
   		return data[index];
 	}
 	T* begin() { return data; }
 	T* end() { return data + count; }
-	void insert(int index, const T& item) {
+	void insert(size_t index, const T& item) {
 	  fb_assert(index >= 0 && index <= count);
 	  fb_assert(count < Capacity);
 	  memmove(data + index + 1, data + index, sizeof(T) * (count++ - index));
 	  data[index] = item;
 	}
 
-	int add(const T& item) {
+	size_t add(const T& item) {
 		fb_assert(count < Capacity);
 		data[count++] = item;
   		return count;
 	};
-	void remove(int index) {
+	void remove(size_t index) {
   		fb_assert(index >= 0 && index < count);
   		memmove(data + index, data + index + 1, sizeof(T) * (--count - index));
 	}
-	void shrink(int newCount) {
+	void shrink(size_t newCount) {
 		fb_assert(newCount <= count);
 		count = newCount;
 	};
@@ -73,10 +73,10 @@ public:
 		memcpy(data + count, L.data, sizeof(T) * L.count);
 		count += L.count;
 	}
-	int getCount() const { return count; }
-	int getCapacity() const { return Capacity; }
+	size_t getCount() const { return count; }
+	size_t getCapacity() const { return Capacity; }
 protected:
-	int count;
+	size_t count;
 	T data[Capacity];
 };
 
@@ -98,16 +98,16 @@ public:
 
 // Fast sorted array of simple objects
 // It is used for B+ tree nodes lower, but can still be used by itself
-template <typename Value, int Capacity, typename Key = Value, 
+template <typename Value, size_t Capacity, typename Key = Value, 
 	typename KeyOfValue = DefaultKeyValue<Value>, 
 	typename Cmp = DefaultComparator<Key> >
 class SortedVector : public Vector<Value, Capacity> {
 public:
 	SortedVector() : Vector<Value, Capacity>() {}
-	bool find(const Key& item, int& pos) const {
-		int highBound = this->count, lowBound = 0;
+	bool find(const Key& item, size_t& pos) const {
+		size_t highBound = this->count, lowBound = 0;
 		while (highBound > lowBound) {
-			const int temp = (highBound + lowBound) >> 1;
+			const size_t temp = (highBound + lowBound) >> 1;
 			if (Cmp::greaterThan(item, KeyOfValue::generate(this, this->data[temp])))
 				lowBound = temp + 1;
 			else
@@ -117,8 +117,8 @@ public:
 		return highBound != this->count &&
 			!Cmp::greaterThan(KeyOfValue::generate(this, this->data[lowBound]), item);
 	}
-	int add(const Value& item) {
-	    int pos;
+	size_t add(const Value& item) {
+	    size_t pos;
   	    find(KeyOfValue::generate(this, item), pos);
 		insert(pos, item);
 		return pos;
